requirements.txt
flask==3.0.2
google-generativeai==0.8.5
python-dotenv==1.0.1
